requirements.txt
fastapi
uvicorn
pydantic-settings
anthropic
openai
google-genai
chromadb
pypdf
